Cleanup - (Task #361)
---
Task #361: Use easiest synthesier per module
