+++ de.dlr.sc.virsat.model.extension.fdir.test/src/de/dlr/sc/virsat/model/extension/fdir/test/AllTests.java
@@ -38,8 +38,9 @@ import de.dlr.sc.virsat.model.extension.fdir.recovery.minimizer.ComposedMinimize
 import de.dlr.sc.virsat.model.extension.fdir.recovery.minimizer.FinalStateMinimizerTest;
 import de.dlr.sc.virsat.model.extension.fdir.recovery.minimizer.OrthogonalPartitionRefinementMinimizerTest;
 import de.dlr.sc.virsat.model.extension.fdir.recovery.minimizer.PartitionRefinementMinimizerTest;
-import de.dlr.sc.virsat.model.extension.fdir.synthesizer.ModularSynthesizerTest;
+import de.dlr.sc.virsat.model.extension.fdir.synthesizer.BasicSynthesizerTest;
 import de.dlr.sc.virsat.model.extension.fdir.synthesizer.DelegateSynthesizerTest;
+import de.dlr.sc.virsat.model.extension.fdir.synthesizer.ModularSynthesizerTest;
 import de.dlr.sc.virsat.model.extension.fdir.synthesizer.POSynthesizerTest;
 import de.dlr.sc.virsat.model.extension.fdir.synthesizer.Schedule2RAConverterTest;
 import de.dlr.sc.virsat.model.extension.fdir.util.FaultTreeHolderTest;
@@ -60,9 +61,10 @@ import junit.framework.JUnit4TestAdapter;
 		DFTStateTest.class,
 		FaultTreeSymmetryCheckerTest.class,
 		GalileoDFT2DFTTest.class,
-		ModularSynthesizerTest.class,
+		BasicSynthesizerTest.class,
 		POSynthesizerTest.class,
 		DelegateSynthesizerTest.class,
+		ModularSynthesizerTest.class,
 		Schedule2RAConverterTest.class,
 		DFT2GalileoDFTTest.class,
 		OrthogonalPartitionRefinementMinimizerTest.class,
